Modify hardware to multiplex data bus controls
This greatly increases the number of available control signals without changing the eeprom word size (or in real life, without changing the number of eeproms).

This works on the observation that at any given time, only one cpu component will be outputting to the data bus. The same goes for reading from the data bus.